requirements.txt
# Python dependencies for Reflective Vest Detection Project

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0

# Data processing
scipy>=1.7.0
pyyaml>=6.0

# Training & Logging
tensorboard>=2.11.0
tqdm>=4.64.0

# Visualization
matplotlib>=3.5.0

# Optional but recommended
scikit-learn>=1.0.0
